r/storage_share: moving the validate function to the validate package
