+++ azurerm/internal/services/storage/validate/storage_share.go
@@ -1,18 +1,19 @@
-package storage
+package validate
 
 import (
 	"fmt"
 	"regexp"
 )
 
-// Following the naming convention as laid out in the docs https://msdn.microsoft.com/library/azure/dn167011.aspx
-func ValidateArmStorageShareName(v interface{}, k string) (warnings []string, errors []error) {
+func ShareName(v interface{}, k string) (warnings []string, errors []error) {
 	value := v.(string)
 	if !regexp.MustCompile(`^[0-9a-z-]+$`).MatchString(value) {
 		errors = append(errors, fmt.Errorf(
 			"only lowercase alphanumeric characters and hyphens allowed in %q: %q",
 			k, value))
 	}
+
+	// Following the naming convention as laid out in the docs https://msdn.microsoft.com/library/azure/dn167011.aspx
 	if len(value) < 3 || len(value) > 63 {
 		errors = append(errors, fmt.Errorf(
 			"%q must be between 3 and 63 characters: %q", k, value))
