+++ azurerm/internal/services/storage/validate/storage_share_test.go
@@ -1,0 +1,34 @@
+package validate
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestValidateShareName(t *testing.T) {
+	validNames := []string{
+		"valid-name",
+		"valid02-name",
+	}
+	for _, v := range validNames {
+		_, errors := ShareName(v, "name")
+		if len(errors) != 0 {
+			t.Fatalf("%q should be a valid Share Name: %q", v, errors)
+		}
+	}
+
+	invalidNames := []string{
+		"InvalidName1",
+		"-invalidname1",
+		"invalid_name",
+		"invalid!",
+		"double-hyphen--invalid",
+		"ww",
+		strings.Repeat("w", 65),
+	}
+	for _, v := range invalidNames {
+		if _, errors := ShareName(v, "name"); len(errors) == 0 {
+			t.Fatalf("%q should be an invalid Share Name", v)
+		}
+	}
+}
